Add stitches column to pattern_thread. This will be a faster way of retriving stitches of projects and will make it easier to create delete functionality.
Currently in order to delete user_stitches, you must know the stitch id. This is very inefficient, so only saving the x/y/stitchedAt so that lookup logic is done in-memory and so they can be deleted by their user_thread id, or at least something more efficient to select.